graph_rag_neo4j/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
neo4j==5.14.0
langchain==0.0.350
langchain-openai==0.0.2.post1
python-multipart==0.0.6
pydantic==2.5.2
tiktoken==0.5.1 